fixed some minor bugs and added ft_striteri

+++ srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: sajabran <sajabran@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/09/13 15:50:31 by sajabran          #+#    #+#             */
-/*   Updated: 2026/09/19 19:21:28 by sajabran         ###   ########.fr       */
+/*   Updated: 2026/09/20 18:07:26 by sajabran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
